Move dhcp6client to pkg directory.
