+++ pkg/dhcp6client/packet.go
